+++ apps/webapp/app/services/externalApis/integrationCatalog.server.ts
@@ -1,11 +1,12 @@
+import { airtable } from "./integrations/airtable";
 import { github } from "./integrations/github";
 import { openai } from "./integrations/openai";
 import { plain } from "./integrations/plain";
 import { resend } from "./integrations/resend";
+import { sendgrid } from "./integrations/sendgrid";
 import { slack } from "./integrations/slack";
 import { stripe } from "./integrations/stripe";
-import { sendgrid } from "./integrations/sendgrid";
-import { supabaseManagement, supabase } from "./integrations/supabase";
+import { supabase, supabaseManagement } from "./integrations/supabase";
 import { typeform } from "./integrations/typeform";
 import type { Integration } from "./types";
 
@@ -30,6 +31,7 @@ export class IntegrationCatalog {
 }
 
 export const integrationCatalog = new IntegrationCatalog({
+  airtable,
   github,
   openai,
   plain,