+++ apps/webapp/app/services/worker.server.ts
@@ -19,6 +19,7 @@ import { DeliverScheduledEventService } from "./schedules/deliverScheduledEvent.
 import { ActivateSourceService } from "./sources/activateSource.server";
 import { DeliverHttpSourceRequestService } from "./sources/deliverHttpSourceRequest.server";
 import { PerformTaskOperationService } from "./tasks/performTaskOperation.server";
+import { addMissingVersionField } from "@trigger.dev/core";
 
 const workerCatalog = {
   indexEndpoint: z.object({
@@ -37,11 +38,21 @@ const workerCatalog = {
     organizationId: z.string(),
     connectionId: z.string(),
   }),
-  activateSource: z.object({
-    id: z.string(),
-    orphanedEvents: z.array(z.string()).optional(),
-  }),
-
+  activateSource: z.preprocess(
+    addMissingVersionField,
+    z.discriminatedUnion("version", [
+      z.object({
+        version: z.literal("1"),
+        id: z.string(),
+        orphanedEvents: z.array(z.string()).optional(),
+      }),
+      z.object({
+        version: z.literal("2"),
+        id: z.string(),
+        orphanedOptions: z.record(z.string(), z.array(z.string())).optional(),
+      }),
+    ])
+  ),
   deliverEvent: z.object({ id: z.string() }),
   "events.invokeDispatcher": z.object({
     id: z.string(),
@@ -187,10 +198,27 @@ function getWorkerQueue() {
       activateSource: {
         priority: 10, // smaller number = higher priority
         maxAttempts: 3,
-        handler: async (payload, job) => {
+        handler: async (payload, graphileJob) => {
           const service = new ActivateSourceService();
-
-          await service.call(payload.id, job.id, payload.orphanedEvents);
+          switch (payload.version) {
+            case "1": {
+              //change the input data to match the new schema
+              await service.call(
+                payload.id,
+                graphileJob.id,
+                payload.orphanedEvents
+                  ? {
+                      event: payload.orphanedEvents,
+                    }
+                  : undefined
+              );
+              break;
+            }
+            case "2": {
+              await service.call(payload.id, graphileJob.id, payload.orphanedOptions);
+              break;
+            }
+          }
         },
       },
       deliverHttpSourceRequest: {
@@ -292,7 +320,12 @@ function getExecutionWorkerQueue() {
         handler: async (payload, job) => {
           const service = new PerformRunExecutionV2Service();
 
-          await service.call(payload.id, payload.reason, payload.isRetry, payload.resumeTaskId);
+          await service.call({
+            id: payload.id,
+            reason: payload.reason,
+            resumeTaskId: payload.resumeTaskId,
+            isRetry: payload.isRetry,
+          });
         },
       },
     },