+++ packages/core/src/schemas/index.ts
@@ -11,3 +11,4 @@ export * from "./notifications";
 export * from "./fetch";
 export * from "./events";
 export * from "./runs";
+export * from "./addMissingVersionField";