+++ packages/database/prisma/migrations/20230904205457_add_max_run_execution_time_to_orgs/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "Organization" ADD COLUMN     "maximumExecutionTimePerRunInMs" INTEGER NOT NULL DEFAULT 900000;